Quasiclassical scattering simulation
*ScatteringSimulation.f90*
Implement the possibility to start a scattering simulation
at 0 K with the oscillator bath and the carbon having
quantum zero point energy. Approximate sampling is done assuming
zero coupling between the carbon and the bath.
The option is available only for normal bath and can be chosen
with the already available input directive ZPECorrection.